fix: type 정리
