+++ src/models/Task.ts
@@ -1,0 +1,4 @@
+export interface Task {
+  id: number;
+  taskName: string;
+}
