+++ src/navigations/types.ts
@@ -1,0 +1,4 @@
+export type RootStackParamList = {
+  Home: React.FC | undefined;
+  AddTask: React.FC | undefined;
+};
